+++ packages/terra-clinical-result/tests/jest/common/other/EnteredInError.test.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import EnteredInError from '../../../../src/common/other/_EnteredInError';
 
 describe('EnteredInError', () => {
   it('should render a default EnteredInError', () => {
-    const result = shallowWithIntl(<EnteredInError />).dive();
+    const result = shallowWithIntl(<EnteredInError />).dive().dive();
     expect(result).toMatchSnapshot();
   });
 });