+++ packages/terra-clinical-result/tests/jest/common/other/KnownNoData.test.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import KnownNoData from '../../../../src/common/other/_KnownNoData';
 
 describe('KnownNoData', () => {
   it('should render a default KnownNoData', () => {
-    const result = shallowWithIntl(<KnownNoData />).dive();
+    const result = shallowWithIntl(<KnownNoData />).dive().dive();
     expect(result).toMatchSnapshot();
   });
 });