+++ packages/terra-clinical-result/tests/jest/common/other/NumericOverflow.test.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import NumericOverflow from '../../../../src/common/other/_NumericOverflow';
 
 describe('NumericOverflow', () => {
   it('should render a default NumericOverflow', () => {
-    const result = shallowWithIntl(<NumericOverflow />).dive();
+    const result = shallowWithIntl(<NumericOverflow />).dive().dive();
     expect(result).toMatchSnapshot();
   });
 });