+++ packages/terra-clinical-result/tests/jest/common/other/ResultError.test.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import ResultError from '../../../../src/common/other/_ResultError';
 
 describe('ResultError', () => {
   it('should render a default ResultError', () => {
-    const result = shallowWithIntl(<ResultError />).dive();
+    const result = shallowWithIntl(<ResultError />).dive().dive();
     expect(result).toMatchSnapshot();
   });
 });