+++ packages/terra-clinical-detail-view/src/terra-dev-site/doc/example/DetailViewDivided.jsx
@@ -9,12 +9,12 @@ const cx = classNames.bind(styles);
 const detail1title = 'Diabetes';
 const detail1item1 = (<LabelValueView label="Age of onset" textValue="38 years" />);
 const detail1item2 = (<LabelValueView label="Reviewed" textValue="07/10/2012 10:00AM" />);
-const detail1item3 = (<LabelValueView label="Comments" textValue="--" />);
+const detail1item3 = (<LabelValueView label="Comments" textValue="Based on patient's blood glucose levels and medical history, I recommend that they continue to monitor their glucose levels regularly and adhere to a healty diet and exercise regimen.It's important for the patient to attend regular follow-up appointments and communicate any concerns in symptoms." />);
 
 const detail2title = 'Hypertension';
-const detail2item1 = (<LabelValueView label="Age of onset" textValue="--" />);
+const detail2item1 = (<LabelValueView label="Age of onset" textValue="38years" />);
 const detail2item2 = (<LabelValueView label="Reviewed" textValue="07/10/2012 10:00AM" />);
-const detail2item3 = (<LabelValueView label="Comments" textValue="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. (Hollister MD, David 09/12/2011 11:34:12AM)" />);
+const detail2item3 = (<LabelValueView label="Comments" textValue="Based on the patient's blood pressure readings and medical history ,I recommend that they make lifestyle modifications such as reducing their salt intake ,increasing physical activity,adn losing weight if necessary.It's important for the patient to continue monitoring their blood pressure at the home and attending regular follow-up appointment to assess the effectiveness of their treatment plan" />);
 
 const DetailViewDivided = () => (
   <div className={cx('detail-view-divided')}>