+++ packages/terra-clinical-header/src/terra-dev-site/doc/example/HeaderLongTextWithContent.jsx
@@ -5,11 +5,11 @@ import Header from 'terra-clinical-header';
 const LongTextWithContent = () => (
   <div>
     <Header
-      startContent={<Button text="Button" />}
-      title="Lorem ipsum dolor sit amet, sea meis iisque petentium an. Vero copiosae eu eos. Usu ocurreret definiebas eu, vim esse duis habemus at, vix at alii ubique. Nobis meliore principes te cum. Ei quando eirmod oportere eum, ei nonumy eruditi eum. In qui odio euripidis, habeo mundi praesent vim et."
-      endContent={<Button text="Button" />}
+      startContent={<Button text="Back" />}
+      title="Patient Information for Individual Diagnosed with a Chronic Disease in the past or currently suffering: Comprehensive Medical History,Treatment Plan,and Prognosis Report"
+      endContent={<Button text="print" />}
     >
-      <Button text="Custom 1" />
+      <Button text="Medical History" />
     </Header>
   </div>
 );