+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated WDIO screenshots due to webpack-config-terra sass upgrade.
+
 ## 3.27.0 - (July 21, 2021)
 
 * Changed