+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Changed
+* Added
   * Added package json lint test.
 
 ## 3.27.0 - (July 21, 2021)