+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Changed
-  * Added `level` prop to allow users to render different heading levels.
+  * Added `level` prop to the DetailView component to allow users to render different heading levels.
 
 ## 3.29.0 - (April 21, 2023)