+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added `level` prop to the DetailView component to allow users to render different heading levels.
+
 ## 3.29.0 - (April 21, 2023)
 
 * Changed