+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -8,7 +8,7 @@
   * Added `level` prop to the DetailView component to allow users to render different heading levels.
 
 * Changed
-  * Update DetailView component to use semantic html elements
+  * Update DetailView component to use semantic html elements.
 
 ## 3.29.0 - (April 21, 2023)